Conflicts:
	src/CMakeLists.txt

+++ src/fe/source.cpp
@@ -2,6 +2,7 @@
 #include "msg/message.h"
 #include <boost/any.hpp>
 #include <vector>
+#include <string>
 
 using namespace psc::fe;
 
@@ -47,7 +48,7 @@ char Source::SourceImpl::next_char()
 
 char Source::SourceImpl::current_char()
 {
-    if (_cpos == -2 || _cpos > _line.size())
+    if (_cpos == -2 || _cpos > static_cast<int>(_line.size()))
     {
         read_line();
         return next_char();
@@ -71,7 +72,7 @@ char Source::SourceImpl::peek_char()
         return END_OF_FILE;
     }
     int npos = _cpos + 1;
-    return npos < _line.size() ? _line[npos] : END_OF_LINE;
+    return npos < static_cast<int>(_line.size()) ? _line[npos] : END_OF_LINE;
 }
 
 // public interface
